etwas aufgeräumt, neues Problem doppelter Buchstabe in Auswahl
